+++ backend/src/main/java/com/backend/category/service/CategoryService.java
@@ -1,6 +1,6 @@
 package com.backend.category.service;
 
-import com.backend.category.dto.CategoryDto;
+import com.backend.category.dto.response.CategoryResponse;
 import com.backend.category.entity.Category;
 import com.backend.category.repository.CategoryRepository;
 import java.util.List;
@@ -15,15 +15,15 @@ public class CategoryService {
     private final CategoryRepository categoryRepository;
 
     // 카테고리 전체 조회
-    public List<CategoryDto> categoryList() {
+    public List<CategoryResponse> categoryList() {
         List<Category> categoryList = categoryRepository.findAll();
         return mappingCategory(categoryList);
     }
 
     // 카테고리 매핑
-    public List<CategoryDto> mappingCategory(List<Category> categoryList) {
+    public List<CategoryResponse> mappingCategory(List<Category> categoryList) {
         return categoryList.stream()
-                .map(category -> CategoryDto.builder()
+                .map(category -> CategoryResponse.builder()
                         .id(category.getId())
                         .name(category.getName())
                         .createdAt(category.getCreatedAt().toLocalDateTime())
